remove v2 designation, delete old code
