+++ src/martok/MartokConfig.ts
@@ -1,0 +1,6 @@
+export type MartokConfig = {
+  package: string;
+  files: string[];
+  output: string;
+  sourceRoot: string;
+};
