+++ src/martok/declarations/DeclarationGenerator.ts
@@ -1,4 +1,4 @@
-import { MartokV2 } from "../MartokV2";
+import { Martok } from "../Martok";
 import {
   isInterfaceDeclaration,
   isTypeAliasDeclaration,
@@ -10,7 +10,7 @@ import { TypeAliasGenerator } from "./TypeAliasGenerator";
 export class DeclarationGenerator {
   private readonly interfaces = new InterfaceGenerator(this.martok);
   private readonly types = new TypeAliasGenerator(this.martok);
-  public constructor(private readonly martok: MartokV2) {}
+  public constructor(private readonly martok: Martok) {}
 
   public generateDeclarations(file: SourceFile): string[] {
     const result: string[] = [];
