--- src/types/MartokFile.ts
@@ -1,7 +1,0 @@
-import { MartokClass } from "./MartokClass";
-
-export type MartokFile = {
-  name: string;
-  package: string;
-  classes: MartokClass[];
-};
