--- src/types/MartokOutput.ts
@@ -1,6 +1,0 @@
-import { MartokFile } from "./MartokFile";
-
-export type MartokOutput = {
-  package: string;
-  files: MartokFile[];
-};
