--- src/types/MartokProperty.ts
@@ -1,7 +1,0 @@
-import { MartokType } from "./MartokType";
-
-export type MartokProperty = {
-  name: string;
-  optional: boolean;
-  type: MartokType;
-};
