--- src/types/MartokType.ts
@@ -1,8 +1,0 @@
-import { SourceFile, Type } from "typescript";
-
-export type MartokType = {
-  name: string;
-  file?: SourceFile;
-  isIntrinsic: boolean;
-  rawType: Type;
-};
